test: moving helper func to os agnostic file
