+++ tsuru/client/archiver_unix_test.go
@@ -10,13 +10,10 @@ package client
 import (
 	"archive/tar"
 	"bytes"
-	"compress/gzip"
-	"errors"
 	"io"
 	"net"
 	"os"
 	"path/filepath"
-	"testing"
 
 	check "gopkg.in/check.v1"
 )
@@ -166,48 +163,4 @@ func (s *S) TestArchive_FilesOnly_MultipleDirs(c *check.C) {
 	c.Assert(stderr.String(), check.Matches, `(?s)(.*)Skipping file "testdata/deploy2/directory/file.txt" as it already exists in the current directory.(.*)`)
 	c.Assert(stderr.String(), check.Matches, `(?s)(.*)Skipping file "testdata/deploy2/file1.txt" as it already exists in the current directory.(.*)`)
 	c.Assert(stderr.String(), check.Matches, `(?s)(.*)Skipping file "testdata/deploy2/file2.txt" as it already exists in the current directory.(.*)`)
-}
-
-func extractFiles(t *testing.T, c *check.C, r io.Reader) (m []miniFile) {
-	t.Helper()
-
-	gzr, err := gzip.NewReader(r)
-	c.Assert(err, check.IsNil)
-
-	tr := tar.NewReader(gzr)
-
-	for {
-		h, err := tr.Next()
-		if errors.Is(err, io.EOF) {
-			break
-		}
-
-		c.Assert(err, check.IsNil)
-
-		var data []byte
-
-		if h.Typeflag == tar.TypeReg || h.Typeflag == tar.TypeRegA {
-			var b bytes.Buffer
-			written, err := io.CopyN(&b, tr, h.Size)
-			c.Assert(err, check.IsNil)
-			c.Assert(written, check.Equals, h.Size)
-			data = b.Bytes()
-		}
-
-		m = append(m, miniFile{
-			Name:     h.Name,
-			Linkname: h.Linkname,
-			Type:     h.Typeflag,
-			Data:     data,
-		})
-	}
-
-	return m
-}
-
-type miniFile struct {
-	Name     string
-	Linkname string
-	Type     byte
-	Data     []byte
 }
